+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Environment section added. Allows adding settings which change the environment.
+- Environment option "enabled". Enables environment configuration.
+- Environment option "nomerges". Allows specifying "nomerges=0,1,2" for devices.
+- Output option "append_environment". Appends environment settings to output.
+
+### Fixed
+- Template settings should appear in output if specified in format.
 
 ## [1.0.0] - 2019-02-26
 ### Changed